dont register extra repl commands for now
